add: refactored actions.ts to be separated into controllers
